chore: align lint config and centralize state vector type
