+++ web/eslint.config.js
@@ -1,14 +1,16 @@
 import js from '@eslint/js'
 import globals from 'globals'
+import react from 'eslint-plugin-react'
 import reactHooks from 'eslint-plugin-react-hooks'
 import reactRefresh from 'eslint-plugin-react-refresh'
-import react from 'eslint-plugin-react'
 import tseslint from 'typescript-eslint'
+
+const reactRecommended = react.configs.flat.recommended
+const reactHooksRecommended = reactHooks.configs.recommended
 
 export default tseslint.config(
   { ignores: ['dist'] },
   {
-    extends: [js.configs.recommended, ...tseslint.configs.recommended],
     files: ['**/*.{ts,tsx}'],
     languageOptions: {
       ecmaVersion: 2020,
@@ -19,9 +21,13 @@ export default tseslint.config(
       'react-hooks': reactHooks,
       'react-refresh': reactRefresh
     },
+    extends: [
+      js.configs.recommended,
+      ...tseslint.configs.recommended,
+      reactRecommended,
+      reactHooksRecommended
+    ],
     rules: {
-      ...react.configs.recommended.rules,
-      ...reactHooks.configs.recommended.rules,
       'react-refresh/only-export-components': 'warn',
       'react/react-in-jsx-scope': 'off',
       'react/prop-types': 'off'
